Fixed a problem with 3 uninitialized variables.
Now the pink ellipse is correctly displayed over the pupil.
Moreover the application can now be calibrated an used, which was not the case
before, since the 'blobs' where rejected by some ifs that were controlled by
the uninitialized variables that happened to have the value 'true'.

Notes:
I originally identified and fixed the problem on 7/7/2013 11:31:12 pm, and
improved the solution (moved the initializations in the corresponding 'setup'
methods) on 11/7/2013 10:54:46 pm. Now I'm just cleaning up the commits for
pushing.

The 'setup' methods should probably be replaced by constructors, but this seems
to be a problem of the application in general and will require major refactoring.