Temporary fix for a portability problem encountered with some webcams.
Now the application uses the same resolution values to scale the pink ellipse
that it uses to configure the webcam, instead of using hard-coded values.
That allows the application to work correctly with webcams that have resolutions
other than 640x480.

Note: I originally fixed this problem on 8/7/2013 12:31:40 am. Now I'm cleaning
up for pushing.